Add tests ensuring that import still works even for unfamiliar filenames (#96)
When importing files, there are two different paths we use for determining the file format.

The first is to pattern match on the filename.

If that fails, we fall back to examining the structure of the file, which can be slower and therefore wasteful.

Before this PR, we only tests that the filename pattern matching was correctly identifying the format. This PR ensures that our file structure matching is working correctly too.