Use `frame.name?.startsWith` for stackprof
Sometimes, stackprof frames don't get generated with a `name` in the
frame. I think it's probably worth tracking down why that is, but in the
mean time, speedscope simply crashes with a method call on `undefined`.
The crash is bad because it only shows up in the console--there's no
visible message saying that speedscope failed to parse and load a
profile.

For more information, see #378

This fixes the crash by simply skipping the logic in demangle if the
name field isn't present on a frame. That's probably a fine tradeoff?
Because in this case, stackprof is generating ruby frames, which means
that C++ name demangling won't apply.

I have tested this by running the scripts/prepare-test-installation.sh
script and verifying that `bin/cli.js` can now successfully load the
included profile. Before these changes, I verified that speedscope
failed with the behavior mentioned in #378.